Refactor (Anizium): Restructure presence and metadata
This commit refactors the Anizium website integration by restructuring the presence and metadata handling. It removes the `poster-manager.ts`, `route-handlers.ts`, and `settings-manager.ts` files, consolidating their logic into new `handlers` and `managers` directories. Additionally, it updates `metadata.json` to include a new logo option "Animated 2" and modifies `presence.ts` to reflect these structural changes.
