+++ websites/A/Anizium/metadata.json
@@ -49,6 +49,7 @@
       "value": 0,
       "values": [
         "Animated",
+        "Animated 2",
         "Still",
         "Still, No Background"
       ],
